% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_ldat.R
\name{clone.ldat}
\alias{clone.ldat}
\title{Clone an ldat object}
\usage{
\method{clone}{ldat}(x, ...)
}
\arguments{
\item{x}{\code{\link{ldat}} object to be cloned}

\item{...}{ignored.}
}
\description{
Clone an ldat object
}
\details{
Clones each of the vectors in the \code{\link{ldat}} object.
}
