% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.R
\name{duplicated.lvec}
\alias{duplicated.lvec}
\alias{unique.lvec}
\title{Find duplicates in a vector}
\usage{
\method{duplicated}{lvec}(x, incomparables = FALSE, fromLast = FALSE, ...)

\method{unique}{lvec}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{an object of type \code{\link{lvec}}.}

\item{incomparables}{passed on to \code{link{duplicated}}.}

\item{fromLast}{not supported.}

\item{...}{passed on to \code{\link{duplicated}}.}
}
\description{
Find duplicates in a vector
}
\details{
Because this function works on chunks of data the data first needs to be 
sorted. Since a non stable sort is used, which of the duplicates is marked
as duplicate is undefined. This is unlike the regular 
\code{\link{duplicated}} in which \code{fromLast} determines which records
are marked as duplicates.

The function processes the data in chunks. The size of the chunks can be 
controlled using the option `chunk_size` (see \code{\link{chunk}}).
}
