% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{[.lvec}
\alias{[.lvec}
\alias{[<-.lvec}
\title{Indexing of lvec objects}
\usage{
\method{[}{lvec}(x, i = NULL, range = NULL)

\method{[}{lvec}(x, i, range) <- value
}
\arguments{
\item{x}{an object of type \code{\link{lvec}}}

\item{i}{an index vector. See \code{\link{lget}}.}

\item{range}{an range of indices. See \code{\link{lget}}.}

\item{value}{new values. See \code{\link{lget}}.}
}
\description{
Indexing of lvec objects
}
\details{
These functions are a wrapper around \code{\link{lget}} and 
\code{\link{lset}}.
}
