% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise_functions.R
\name{is.na.lvec}
\alias{is.na.lvec}
\title{Simple elementwise functions}
\usage{
\method{is.na}{lvec}(x)
}
\arguments{
\item{x}{an object f type \code{\link{lvec}}}
}
\value{
Returns an \code{\link{lvec}} of the same length as the input.
}
\description{
These are implementations for \code{\link{lvec}} object for their regular
R counterparts.
}
