% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lset_ldat.R
\name{lset.ldat}
\alias{lset.ldat}
\title{Set values is an ldat object}
\usage{
\method{lset}{ldat}(x, index = NULL, values, range = NULL, ...)
}
\arguments{
\item{x}{an object of type \code{\link{ldat}}}

\item{index}{a numeric of logical vector with indices at which 
the values should be set.}

\item{values}{a vector with new values.}

\item{range}{a numeric vector of length 2 specifying a range of elements 
to select. Specify either \code{index} or \code{range}.}

\item{...}{ignored.}
}
\description{
Set values is an ldat object
}
\details{
When values is a vector the values are assigned to each column in
\code{x}. Otherwise, vectors is assumed to be a list or data.frame
of the same length as \code{x}. Each element of \code{values} is 
assigned to the corresponding element of \code{x}.
}
