% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{Math.lvec}
\alias{Math.lvec}
\title{Implementation of Math group generics for lvec}
\usage{
\method{Math}{lvec}(x, ...)
}
\arguments{
\item{x}{an object of type \code{\link{lvec}}.}

\item{...}{passed on to the corresponding R functions}
}
\value{
Returns an \code{link{lvec}} of the same length as the input.
}
\description{
Implementation of Math group generics for lvec
}
\details{
Math is group generic implementing the following functions: 
\code{\link{abs}}, \code{\link{sign}}, \code{\link{sqrt}},
\code{\link{floor}}, \code{\link{ceiling}}, \code{\link{trunc}},
\code{\link{round}}, \code{\link{signif}} \code{\link{exp}}, 
\code{\link{log}}, \code{\link{expm1}}, \code{\link{log1p}},
\code{\link{cos}}, \code{\link{sin}}, \code{\link{tan}},
\code{\link{cospi}}, \code{\link{sinpi}}, \code{\link{tanpi}},
\code{\link{acos}}, \code{\link{asin}}, \code{\link{atan}},
\code{\link{cosh}}, \code{\link{sinh}}, \code{\link{tanh}},
\code{\link{acosh}}, \code{\link{asinh}}, \code{\link{atanh}},
\code{\link{lgamma}}, \code{\link{gamma}}, \code{\link{digamma}}, 
\code{\link{trigamma}}, \code{\link{cumsum}}, \code{\link{cumprod}}, 
\code{\link{cummax}}, \code{\link{cummin}}. For more information
see \code{\link{Math}}.
}
