% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_sort.R
\name{median.lvec}
\alias{median.lvec}
\title{Calculate the median of an lvec}
\usage{
\method{median}{lvec}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{an object of type \code{\link{lvec}}.}

\item{na.rm}{remove missing values before calculating the quantiles}

\item{...}{ignored.}
}
\description{
Calculate the median of an lvec
}
\seealso{
For more details see \code{\link{quantile.lvec}}.
}
