% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Ops.lvec}
\alias{Ops.lvec}
\title{Implementation of Ops group generics for lvec}
\usage{
\method{Ops}{lvec}(e1, e2)
}
\arguments{
\item{e1}{an object of type \code{\link{lvec}}.}

\item{e2}{an object of type \code{\link{lvec}}.}
}
\value{
Returns an \code{link{lvec}} of the same length as the input.
}
\description{
Implementation of Ops group generics for lvec
}
\details{
Math is group generic implementing the following functions: 
‘"+"’, ‘"-"’, ‘"*"’, ‘"/"’, ‘"^"’, ‘"%%"’, ‘"%/%"’ ‘"&"’, 
‘"|"’, ‘"!"’ ‘"=="’, ‘"!="’, ‘"<"’, ‘"<="’, ‘">="’, ‘">"’.
For more information see \code{\link{Ops}}.
}
