% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_summary.R
\name{all.lvec}
\alias{all.lvec}
\alias{any.lvec}
\alias{prod.lvec}
\alias{sum.lvec}
\alias{mean.lvec}
\alias{max.lvec}
\alias{min.lvec}
\alias{range.lvec}
\title{Basic summary functions for lvec objects}
\usage{
\method{all}{lvec}(x, ..., na.rm = FALSE)

\method{any}{lvec}(x, ..., na.rm = FALSE)

\method{prod}{lvec}(x, ..., na.rm = FALSE)

\method{sum}{lvec}(x, ..., na.rm = FALSE)

\method{mean}{lvec}(x, ..., na.rm = FALSE)

\method{max}{lvec}(x, ..., na.rm = FALSE)

\method{min}{lvec}(x, ..., na.rm = FALSE)

\method{range}{lvec}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{an \code{\link{lvec}} object}

\item{...}{ignored.}

\item{na.rm}{logical indicating whether missing values should be ignored}
}
\description{
These functions should behave as their regular counterparts.
}
