% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_ldat.R
\name{order.ldat}
\alias{order.ldat}
\title{Order an ldat}
\usage{
\method{order}{ldat}(x, ...)
}
\arguments{
\item{x}{\code{\link{ldat}} to sort}

\item{...}{unused.}
}
\value{
Returns the order of \code{x}. Unlike the default \code{\link{order}} 
function in R, the sort used is not stable (e.g. in case there are multiple
records with the same value in \code{x}, there relative order after sorting
is not defined).
}
\description{
Order an ldat
}
\examples{
x <- as_ldat(iris)
o <- order(x[c("Sepal.Width", "Sepal.Length")])

}
