% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_ldat.R
\name{sort.ldat}
\alias{sort.ldat}
\title{Sort an ldat}
\usage{
\method{sort}{ldat}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{\code{\link{ldat}} to sort}

\item{decreasing}{unused (a value unequal to \code{FALSE} will generate an error).}

\item{...}{unused.}
}
\value{
Sorts \code{x} and returns a sorted copy of \code{x}.
}
\description{
Sort an ldat
}
\examples{
x <- as_ldat(iris)
sort(x)

}
