% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table.lvec}
\alias{table.lvec}
\title{Create cross tables from lvec objects}
\usage{
\method{table}{lvec}(..., useNA = c("ifany", "no", "always"),
  chunk_size = 1e+06)
}
\arguments{
\item{...}{an object of type \code{\link{lvec}}}

\item{useNA}{what to do with missing values. See \code{\link{table}}.}

\item{chunk_size}{an integer specifying the size of the chunks with which the
\code{\link{lvec}} objects are processed.}
}
\description{
Create cross tables from lvec objects
}
\seealso{
This function duplicates the functionalty of the \code{\link{table}} 
function.
}
