% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which}
\alias{which}
\alias{which.default}
\alias{which.lvec}
\title{Give the `TRUE` indices of an lvec}
\usage{
which(x, ...)

\method{which}{default}(x, ...)

\method{which}{lvec}(x, ...)
}
\arguments{
\item{x}{logical \code{\link{lvec}} to get the indices from}

\item{...}{not used}
}
\value{
Returns a numeric lvec with the indices of the elements of \code{x} that are 
TRUE.
}
\description{
Give the `TRUE` indices of an lvec
}
\examples{
x <- as_lvec(runif(1E6) > 0.1)
which(x)

}

