% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{[.lvec}
\alias{[.lvec}
\alias{[<-.lvec}
\alias{[.ldat}
\alias{[<-.ldat}
\title{Indexing of lvec objects}
\usage{
\method{[}{lvec}(x, i = NULL, range = NULL)

\method{[}{lvec}(x, i, range) <- value

\method{[}{ldat}(x, i, j, drop = FALSE, range = NULL, clone = TRUE)

\method{[}{ldat}(x, i, range) <- value
}
\arguments{
\item{x}{an object of type \code{\link{lvec}}}

\item{i}{an index vector. See \code{\link{lget}}.}

\item{range}{an range of indices. See \code{\link{lget}}.}

\item{value}{new values. See \code{\link{lget}}.}

\item{j}{a selection of columns (a character, numeric or logical vector).}

\item{drop}{ignored; included for compatability with \code{data.frame}.}

\item{clone}{\code{\link{clone}} columns when selecting only columns.}
}
\description{
Indexing of lvec objects
}
\details{
These functions are a wrapper around \code{\link{lget}} and 
\code{\link{lset}}.
}
